fix: move keycloak client to utils
