+++ config/utils/keycloak_client.go
@@ -1,4 +1,4 @@
-package clients
+package utils
 
 import (
 	"context"
